Create workaround for when json response changes.
ok, i'm going to sound ca-razy right now, but i swear this
happened: in the middle of working on this the json response
changed from providing a customer and media objects to just
the media objects... and the pictures were mostly the same
EXCEPT for one of a dude with a plunger.  my (obvious)
interpretation is that luigi from super mario bros. is stuck
in the internet and signaling me for help.

more on this as it develops.